+++ src/functionalTest/java/uk/gov/hmcts/reform/em/hrs/functional/DownloadHearingRecordingScenarios.java
@@ -1,16 +1,12 @@
-package uk.gov.hmcts.reform.em.hrs;
+package uk.gov.hmcts.reform.em.hrs.functional;
 
 import com.fasterxml.jackson.databind.JsonNode;
-import io.restassured.response.Response;
-import org.apache.commons.lang3.RandomUtils;
 import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
-import org.junit.jupiter.api.DisplayName;
 import org.springframework.beans.factory.annotation.Autowired;
-import uk.gov.hmcts.reform.ccd.client.model.CallbackRequest;
 import uk.gov.hmcts.reform.ccd.client.model.CaseDetails;
-import uk.gov.hmcts.reform.em.hrs.testutil.TestUtil;
+import uk.gov.hmcts.reform.em.hrs.functional.util.TestUtil;
 
 import java.util.Optional;
 import java.util.UUID;
@@ -22,7 +18,7 @@ import static org.hamcrest.Matchers.not;
 import static org.junit.Assert.assertNotNull;
 import static org.junit.Assert.assertTrue;
 
-public class ShareHearingRecordingScenarios extends BaseTest {
+public class DownloadHearingRecordingScenarios extends BaseTest {
 
     @Autowired
     private TestUtil testUtil;
@@ -39,14 +35,13 @@ public class ShareHearingRecordingScenarios extends BaseTest {
     }
 
     @After
-    public void clear(){
+    public void clear() {
         testUtil.deleteFileFromHrsContainer(FOLDER);
         testUtil.deleteFileFromCvpContainer(FOLDER);
     }
 
     @Test
-    @DisplayName("Sharee with caseworker role should be able to download hearing recordings")
-    public void shouldAbleToShareHearingRecordingsToEmailAddressAndDownload() throws Exception {
+    public void anUserWithCaseWorkerHrsRoleShouldBeAbleToDownloadHearingRecordings() throws Exception {
         final JsonNode segmentPayload = getSegmentPayload(FILE_NAME);
 
         postRecordingSegment(segmentPayload)
@@ -64,34 +59,23 @@ public class ShareHearingRecordingScenarios extends BaseTest {
         assertNotNull(caseDetails.getId());
         assertNotNull(caseDetails.getData());
 
-        final CallbackRequest callbackRequest = getCallbackRequest(caseDetails, SHAREE_EMAIL_ADDRESS);
-        final Response shareRecordingResponse = shareRecording(SHAREE_EMAIL_ADDRESS, CASE_WORKER_ROLE, callbackRequest);
-
-        shareRecordingResponse
-            .then()
-            .log().all()
-            .assertThat()
-            .statusCode(200);
-
         final int expectedFileSize = testUtil.getTestFile().readAllBytes().length;
         assertThat(expectedFileSize, is(not(0)));
 
         final byte[] downloadedFileBytes =
-            downloadRecording(caseDetails.getData())
+            downloadRecording(EMAIL_ADDRESS, CASE_WORKER_HRS_ROLE, caseDetails.getData())
                 .then()
                 .statusCode(200)
                 .extract().response()
                 .body().asByteArray();
 
         final int actualFileSize = downloadedFileBytes.length;
-
         assertThat(actualFileSize, is(not(0)));
         assertThat(actualFileSize, is(expectedFileSize));
     }
 
     @Test
-    @DisplayName("Sharee with citizen role should not be able to download a hearing recording")
-    public void shareeWithCitizenRoleShouldNotBeAbleToDownloadHearingRecordings() throws Exception {
+    public void anUserWithCaseWorkerRoleShouldNotBeAbleToDownloadHearingRecordings() throws Exception {
         final JsonNode segmentPayload = getSegmentPayload(FILE_NAME);
 
         postRecordingSegment(segmentPayload)
@@ -109,85 +93,53 @@ public class ShareHearingRecordingScenarios extends BaseTest {
         assertNotNull(caseDetails.getId());
         assertNotNull(caseDetails.getData());
 
-        final CallbackRequest callbackRequest = getCallbackRequest(caseDetails, SHAREE_EMAIL_ADDRESS);
-        final Response shareRecordingResponse = shareRecording(SHAREE_EMAIL_ADDRESS, CITIZEN_ROLE, callbackRequest);
+        final int expectedFileSize = testUtil.getTestFile().readAllBytes().length;
+        assertThat(expectedFileSize, is(not(0)));
 
-        shareRecordingResponse
+        final byte[] downloadedFileBytes =
+            downloadRecording(EMAIL_ADDRESS, CASE_WORKER_ROLE, caseDetails.getData())
+                .then()
+                .statusCode(200) //FIXME should return 403
+                .extract().response()
+                .body().asByteArray();
+
+        final int actualFileSize = downloadedFileBytes.length;
+        assertThat(actualFileSize, is(not(0)));
+        assertThat(actualFileSize, is(expectedFileSize));
+
+    }
+
+    @Test
+    public void anUserWithCitizenRoleShouldNotBeAbleToDownloadHearingRecordings() throws Exception {
+        final JsonNode segmentPayload = getSegmentPayload(FILE_NAME);
+
+        postRecordingSegment(segmentPayload)
             .then()
             .log().all()
-            .assertThat()
-            .statusCode(200);
+            .statusCode(202);
+
+        TimeUnit.SECONDS.sleep(20);
+
+        final Optional<CaseDetails> optionalCaseDetails = findCaseDetailsInCCDByRecordingReference(FILE_NAME);
+        assertTrue(optionalCaseDetails.isPresent());
+
+        final CaseDetails caseDetails = optionalCaseDetails.orElseGet(() -> CaseDetails.builder().build());
+        assertNotNull(caseDetails);
+        assertNotNull(caseDetails.getId());
+        assertNotNull(caseDetails.getData());
 
         final int expectedFileSize = testUtil.getTestFile().readAllBytes().length;
         assertThat(expectedFileSize, is(not(0)));
 
         final byte[] downloadedFileBytes =
-            downloadRecording(SHAREE_EMAIL_ADDRESS, CITIZEN_ROLE, caseDetails.getData())
+            downloadRecording(EMAIL_ADDRESS, CITIZEN_ROLE, caseDetails.getData())
                 .then()
-                .statusCode(200) //FIXME citizen role should not download hearing recordings
+                .statusCode(200) //FIXME should return 403
                 .extract().response()
                 .body().asByteArray();
 
         final int actualFileSize = downloadedFileBytes.length;
-
         assertThat(actualFileSize, is(not(0)));
         assertThat(actualFileSize, is(expectedFileSize));
-    }
-
-    @Test
-    @DisplayName("Should not share a hearing recording to an invalid email address")
-    public void shouldReturn400WhenShareHearingRecordingsToInvalidEmailAddress() throws Exception {
-        final JsonNode segmentPayload = getSegmentPayload(FILE_NAME);
-
-        postRecordingSegment(segmentPayload)
-            .then()
-            .log().all()
-            .statusCode(202);
-
-        TimeUnit.SECONDS.sleep(20);
-
-        final Optional<CaseDetails> optionalCaseDetails = findCaseDetailsInCCDByRecordingReference(FILE_NAME);
-        assertTrue(optionalCaseDetails.isPresent());
-
-        CaseDetails caseDetails = optionalCaseDetails.orElseGet(() -> CaseDetails.builder().build());
-        assertNotNull(caseDetails);
-        assertNotNull(caseDetails.getId());
-        assertNotNull(caseDetails.getData());
-
-        final CallbackRequest callbackRequest = getCallbackRequest(caseDetails, ERROR_SHAREE_EMAIL_ADDRESS);
-        final Response shareRecordingResponse = shareRecording(SHAREE_EMAIL_ADDRESS, CASE_WORKER_ROLE, callbackRequest);
-
-        shareRecordingResponse
-            .then().log().all()
-            .assertThat().statusCode(400);
-    }
-
-    @Test
-    @DisplayName("Should not share a hearing recording when case id does not exist")
-    public void shouldReturn404WhenShareHearingRecordingsToEmailAddressWithNonExistentCaseId() throws Exception {
-        final JsonNode segmentPayload = getSegmentPayload(FILE_NAME);
-
-        postRecordingSegment(segmentPayload)
-            .then()
-            .log().all()
-            .statusCode(202);
-
-        TimeUnit.SECONDS.sleep(20);
-
-        final Optional<CaseDetails> optionalCaseDetails = findCaseDetailsInCCDByRecordingReference(FILE_NAME);
-        assertTrue(optionalCaseDetails.isPresent());
-
-        CaseDetails caseDetails = optionalCaseDetails.orElseGet(() -> CaseDetails.builder().build());
-        assertNotNull(caseDetails);
-        assertNotNull(caseDetails.getId());
-        assertNotNull(caseDetails.getData());
-
-        caseDetails.setId(RandomUtils.nextLong());
-        final CallbackRequest callbackRequest = getCallbackRequest(caseDetails, SHAREE_EMAIL_ADDRESS);
-        final Response shareRecordingResponse = shareRecording(SHAREE_EMAIL_ADDRESS, CASE_WORKER_ROLE, callbackRequest);
-
-        shareRecordingResponse
-            .then().log().all()
-            .assertThat().statusCode(404);
     }
 }
